fix(build): house the DSL matcher server-side — safeRegex's native re2 broke the Vite client build
src/utils/autoResponderUtils.ts is imported by frontend TriggerItem.tsx;
adding homoglyph/safeRegex imports there pulled the native re2 binary
into the client bundle graph (UNLOADABLE_DEPENDENCY). The matcher now
lives in src/server/utils/autoResponderMatcher.ts with a comment
explaining the constraint; autoResponderUtils.ts restored to its
import-free pre-PR state.
